Re/move obsolete JS
