+++ theme/base/javascripts/index.js
@@ -1,4 +1,4 @@
-import {replaceMetadataCertificateLinkTexts} from "./modules/EngineBlockMainPage";
+import {replaceMetadataCertificateLinkTexts} from "./index/EngineBlockMainPage";
 import {initializePage} from './page';
 
 export function initializeIndex() {
