--- theme/base/javascripts/modules/UnconnectedIdpPicker.js
@@ -1,3 +1,0 @@
-import {AbstractIdpPicker} from "./AbstractIdpPicker";
-
-export class UnconnectedIdpPicker extends AbstractIdpPicker {}
