+++ theme/cypress/integration/openconext/wayf/UnconnectedIdpPicker.spec.js
@@ -1,4 +1,4 @@
-import {UnconnectedIdpPicker} from "../../../../base/javascripts/modules/UnconnectedIdpPicker";
+import {UnconnectedIdpPicker} from "../../../../openconext/javascripts/wayf/UnconnectedIdpPicker";
 
 it('Check if the UnconnectedIdpPicker constructor was called', () => {
     const unconnectedIdpPicker = new UnconnectedIdpPicker(null, null, null);
